* Fixes for WTI-Review-587: ** Imported update database design such that FailureMechanismSectionPointEntity now has Order of type int (was long) ** Fixed typos Related to Issue [WTI-575]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@3481 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 76d15ffe3a160eb1c33f2454deaded5a2a3b000d
Former-commit-id: d17ab777c7c5a4b6fa78090a51b57977b7c3e94f

+++ Application/Ringtoets/src/Application.Ringtoets.Storage/DbContext/FailureMechanismSectionPointEntity.cs
@@ -39,7 +39,7 @@ namespace Application.Ringtoets.Storage.DbContext
         public long FailureMechanismSectionEntityId { get; set; }
         public decimal X { get; set; }
         public decimal Y { get; set; }
-        public long Order { get; set; }
+        public int Order { get; set; }
     
         public virtual FailureMechanismSectionEntity FailureMechanismSectionEntity { get; set; }
     }
